refactor(site): consolidate Vite integration
